requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
colorlog==6.9.0
audioop-lts
